LPS-111991 Rename FieldSet files and replace their imports on app-builder
