+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/utils/saveDataDefinition.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {addItem, updateItem} from '../../utils/client.es';
+import {addItem, updateItem} from './client.es';
 
 export default ({
 	dataDefinition,
