+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/index.es.js
@@ -14,7 +14,8 @@
 
 import FormBuilderWithLayoutProvider from './components/FormBuilder/FormBuilderWithLayoutProvider.es';
 import RuleEditor from './components/RuleEditor/RuleEditor.es';
+import * as FieldSupport from './util/fieldSupport.es';
 
 export default FormBuilderWithLayoutProvider;
 
-export {RuleEditor};
+export {FieldSupport, RuleEditor};
